[refactor] 파일 분리
