+++ app/src/main/java/com/flint/presentation/collectiondetail/CollectionDetailScreen.kt
@@ -1,32 +1,19 @@
 package com.flint.presentation.collectiondetail
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.LocalOverscrollFactory
 import androidx.compose.foundation.ScrollState
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.LaunchedEffect
@@ -35,33 +22,19 @@ import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.layout.LayoutCoordinates
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.layout.positionInParent
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.flint.R
-import com.flint.core.common.extension.noRippleClickable
 import com.flint.core.common.util.UiState
-import com.flint.core.designsystem.component.button.FlintSaveDoneButton
-import com.flint.core.designsystem.component.button.FlintSaveNoneButton
 import com.flint.core.designsystem.component.collection.PeopleBottomSheet
-import com.flint.core.designsystem.component.collection.Spoiler
-import com.flint.core.designsystem.component.image.NetworkImage
-import com.flint.core.designsystem.component.image.ProfileImage
 import com.flint.core.designsystem.component.indicator.FlintLoadingIndicator
 import com.flint.core.designsystem.component.progressbar.UnderImageProgressBar
 import com.flint.core.designsystem.component.toast.ShowSaveToast
@@ -73,6 +46,10 @@ import com.flint.domain.model.bookmark.CollectionBookmarkUsersModel
 import com.flint.domain.model.collection.CollectionDetailModelNew
 import com.flint.domain.model.content.ContentModelNew
 import com.flint.domain.type.UserRoleType
+import com.flint.presentation.collectiondetail.component.CollectionDetailDescription
+import com.flint.presentation.collectiondetail.component.CollectionDetailThumbnail
+import com.flint.presentation.collectiondetail.component.Content
+import com.flint.presentation.collectiondetail.component.PeopleWhoSavedThisCollection
 import com.flint.presentation.collectiondetail.sideeffect.CollectionDetailSideEffect
 import com.flint.presentation.collectiondetail.uistate.CollectionDetailUiState
 import kotlinx.collections.immutable.ImmutableList
@@ -109,6 +86,7 @@ fun CollectionDetailRoute(
             CollectionDetailScreen(
                 paddingValues = paddingValues,
                 targetImageUrl = targetImageUrl,
+                thumbnailUrl = collectionDetail.thumbnailUrl,
                 title = collectionDetail.title,
                 isBookmarked = collectionDetail.isBookmarked,
                 authorNickname = collectionDetail.author.nickname,
@@ -215,6 +193,7 @@ fun CollectionDetailRoute(
 @Composable
 fun CollectionDetailScreen(
     paddingValues: PaddingValues,
+    thumbnailUrl: String,
     title: String,
     isBookmarked: Boolean,
     authorNickname: String,
@@ -268,26 +247,29 @@ fun CollectionDetailScreen(
         }
 
         Column(
-            modifier =
-                Modifier
-                    .fillMaxSize()
-                    .padding(paddingValues)
-                    .background(color = FlintTheme.colors.background),
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(paddingValues)
+                .background(color = FlintTheme.colors.background),
         ) {
-            FlintBackTopAppbar(
-                onClick = navigateUp,
-                backgroundColor = Color.Transparent,
-            )
+//            FlintBackTopAppbar(
+//                onClick = navigateUp,
+//                backgroundColor = Color.Transparent,
+//            )
 
             Box {
+                FlintBackTopAppbar(
+                    onClick = navigateUp,
+                    backgroundColor = Color.Transparent,
+                )
                 Column(
-                    modifier =
-                        Modifier
-                            .fillMaxSize()
-                            .verticalScroll(scrollState)
-                            .padding(bottom = 24.dp),
+                    modifier = Modifier
+                        .fillMaxSize()
+                        .verticalScroll(scrollState)
+                        .padding(bottom = 24.dp),
                 ) {
-                    Thumbnail(
+                    CollectionDetailThumbnail(
+                        thumbnailImage = thumbnailUrl,
                         title = title,
                         isBookmarked = isBookmarked,
                         onSaveDoneButtonClick = onSaveDoneButtonClick,
@@ -352,571 +334,8 @@ fun CollectionDetailScreen(
     }
 }
 
-@Composable
-private fun PeopleWhoSavedThisCollection(
-    people: ImmutableList<CollectionBookmarkUsersModel.User>,
-    onMoreClick: () -> Unit,
-) {
-    Column(
-        modifier = Modifier.padding(vertical = 10.dp),
-    ) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Text(
-                text = "이 컬렉션을 저장한 사람들",
-                color = FlintTheme.colors.white,
-                style = FlintTheme.typography.head2Sb20,
-                modifier =
-                    Modifier
-                        .weight(1f)
-                        .padding(horizontal = 16.dp),
-            )
-
-            Icon(
-                imageVector = ImageVector.vectorResource(R.drawable.ic_more),
-                contentDescription = null,
-                modifier =
-                    Modifier
-                        .size(48.dp)
-                        .clickable(onClick = onMoreClick)
-                        .padding(12.dp),
-                tint = FlintTheme.colors.white,
-            )
-        }
-
-        Spacer(Modifier.height(24.dp))
-
-        Row(
-            modifier = Modifier.padding(horizontal = 16.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Row(
-                horizontalArrangement = Arrangement.spacedBy((-12).dp),
-            ) {
-                people.take(5).forEach { author: CollectionBookmarkUsersModel.User ->
-                    ProfileImage(
-                        imageUrl = author.profileImageUrl,
-                        modifier =
-                            Modifier
-                                .size(56.dp)
-                                .border(3.dp, FlintTheme.colors.background, CircleShape),
-                        contentDescription = author.nickName,
-                    )
-                }
-            }
-
-            Spacer(Modifier.width(8.dp))
-
-            if (people.size >= 6) {
-                Row {
-                    Icon(
-                        imageVector = ImageVector.vectorResource(R.drawable.ic_plus),
-                        contentDescription = "그 외",
-                        tint = FlintTheme.colors.white,
-                    )
-
-                    Text(
-                        text = (people.size - 5).toString(),
-                        color = FlintTheme.colors.gray50,
-                        style = FlintTheme.typography.head2M20,
-                    )
-                }
-            }
-        }
-    }
-}
-
-private class PeoplePreviewProvider :
-    PreviewParameterProvider<ImmutableList<CollectionBookmarkUsersModel.User>> {
-    override val values: Sequence<ImmutableList<CollectionBookmarkUsersModel.User>> =
-        sequenceOf(
-            persistentListOf(
-                CollectionBookmarkUsersModel.User(
-                    userId = "1",
-                    nickName = "유저1",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.FLING,
-                ),
-            ),
-            persistentListOf(
-                CollectionBookmarkUsersModel.User(
-                    userId = "1",
-                    nickName = "유저1",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.ADMIN,
-                ),
-                CollectionBookmarkUsersModel.User(
-                    userId = "2",
-                    nickName = "유저2",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.FLINER,
-                ),
-                CollectionBookmarkUsersModel.User(
-                    userId = "3",
-                    nickName = "유저3",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.FLING,
-                ),
-                CollectionBookmarkUsersModel.User(
-                    userId = "4",
-                    nickName = "유저4",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.FLING,
-                ),
-                CollectionBookmarkUsersModel.User(
-                    userId = "5",
-                    nickName = "유저5",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.FLING,
-                ),
-            ),
-            persistentListOf(
-                CollectionBookmarkUsersModel.User(
-                    userId = "1",
-                    nickName = "유저1",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.FLING,
-                ),
-                CollectionBookmarkUsersModel.User(
-                    userId = "2",
-                    nickName = "유저2",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.FLING,
-                ),
-                CollectionBookmarkUsersModel.User(
-                    userId = "3",
-                    nickName = "유저3",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.FLING,
-                ),
-                CollectionBookmarkUsersModel.User(
-                    userId = "4",
-                    nickName = "유저4",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.FLING,
-                ),
-                CollectionBookmarkUsersModel.User(
-                    userId = "5",
-                    nickName = "유저5",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.FLING,
-                ),
-                CollectionBookmarkUsersModel.User(
-                    userId = "6",
-                    nickName = "유저6",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.FLINER,
-                ),
-                CollectionBookmarkUsersModel.User(
-                    userId = "7",
-                    nickName = "유저7",
-                    profileImageUrl = "",
-                    userRole = UserRoleType.ADMIN,
-                ),
-            ),
-        )
-}
-
-@Preview
-@Composable
-private fun PeopleWhoSavedThisCollectionPreview(
-    @PreviewParameter(PeoplePreviewProvider::class) people: ImmutableList<CollectionBookmarkUsersModel.User>,
-) {
-    FlintTheme {
-        PeopleWhoSavedThisCollection(
-            people = people,
-            onMoreClick = {},
-        )
-    }
-}
-
-@Composable
-private fun Thumbnail(
-    title: String,
-    isBookmarked: Boolean,
-    onSaveDoneButtonClick: () -> Unit,
-    onSaveNoneButtonClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        modifier =
-            modifier
-                .fillMaxWidth()
-                .aspectRatio(360f / 270f),
-    ) {
-        Image(
-            painter = painterResource(R.drawable.img_collection_bg2),
-            contentDescription = null,
-            modifier = Modifier.fillMaxSize(),
-            contentScale = ContentScale.FillBounds,
-        )
-
-        Column(
-            modifier =
-                Modifier
-                    .fillMaxSize()
-                    .padding(horizontal = 16.dp)
-                    .padding(top = 55.dp, bottom = 19.dp),
-            horizontalAlignment = Alignment.End,
-            verticalArrangement = Arrangement.SpaceBetween,
-        ) {
-            Text(
-                text = title,
-                color = FlintTheme.colors.white,
-                style = FlintTheme.typography.display2M28,
-                modifier = Modifier.fillMaxWidth(),
-                maxLines = 2,
-                overflow = TextOverflow.Ellipsis
-            )
-            if (isBookmarked) {
-                FlintSaveDoneButton(
-                    onClick = {
-                        onSaveDoneButtonClick()
-                    },
-                )
-            } else {
-                FlintSaveNoneButton(
-                    onClick = {
-                        onSaveNoneButtonClick()
-                    },
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun CollectionDetailDescription(
-    authorNickname: String,
-    authorUserRoleType: UserRoleType,
-    createdAt: String,
-    collectionContent: String,
-    onAuthorNicknameClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(24.dp),
-    ) {
-        Row(
-            horizontalArrangement = Arrangement.spacedBy(10.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Text(
-                text = authorNickname,
-                color = FlintTheme.colors.white,
-                style = FlintTheme.typography.head2Sb20,
-                modifier = Modifier.noRippleClickable(
-                    onClick = { onAuthorNicknameClick() }
-                )
-            )
-
-            if (authorUserRoleType == UserRoleType.FLINER) {
-                Icon(
-                    imageVector = ImageVector.vectorResource(R.drawable.ic_qualified),
-                    contentDescription = "플리너",
-                    tint = Color.Unspecified,
-                )
-            } else {
-                Text(
-                    "|",
-                    color = FlintTheme.colors.gray200,
-                    style = FlintTheme.typography.head3M18,
-                )
-            }
-
-            Text(
-                text = createdAt,
-                color = FlintTheme.colors.gray200,
-                style = FlintTheme.typography.body2M14,
-            )
-        }
-
-        Box(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .height(1.dp)
-                    .clip(CircleShape)
-                    .background(color = FlintTheme.colors.gray300),
-        )
-
-        Text(
-            text = collectionContent,
-            color = FlintTheme.colors.gray100,
-            style = FlintTheme.typography.body1R16,
-        )
-    }
-}
-
-@Composable
-private fun Content(
-    content: ContentModelNew,
-    onBookmarkIconClick: (contentId: String) -> Unit,
-    onSpoilClick: (contentId: String) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Column(modifier = modifier) {
-        NetworkImage(
-            imageUrl = content.imageUrl,
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .aspectRatio(360f / 480f),
-            contentScale = ContentScale.Crop,
-        )
-
-        Spacer(Modifier.height(32.dp))
-
-        Column(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp),
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-            ) {
-                Column {
-                    Text(
-                        content.title,
-                        color = FlintTheme.colors.white,
-                        style = FlintTheme.typography.head2Sb20,
-                    )
-
-                    Spacer(Modifier.height(12.dp))
-
-                    Text(
-                        content.year.toString(),
-                        color = FlintTheme.colors.gray300,
-                        style = FlintTheme.typography.body1R16,
-                    )
-
-                    Text(
-                        content.director,
-                        color = FlintTheme.colors.gray300,
-                        style = FlintTheme.typography.body1R16,
-                    )
-                }
-
-                Row(
-                    modifier =
-                        Modifier.noRippleClickable(onClick = { onBookmarkIconClick(content.id) }),
-                ) {
-                    Column(
-                        modifier =
-                            Modifier
-                                .padding(start = 24.dp)
-                                .padding(vertical = 3.dp),
-                        horizontalAlignment = Alignment.CenterHorizontally,
-                        verticalArrangement = Arrangement.Center,
-                    ) {
-                        if (content.isBookmarked) {
-                            Icon(
-                                imageVector = ImageVector.vectorResource(R.drawable.ic_bookmark_fill),
-                                contentDescription = "저장됨",
-                                tint = Color.Unspecified,
-                            )
-                        } else {
-                            Icon(
-                                imageVector = ImageVector.vectorResource(R.drawable.ic_bookmark_empty),
-                                contentDescription = "저장되지 않음",
-                                tint = Color.White,
-                            )
-                        }
-
-                        Text(
-                            text = content.bookmarkCount.toString(),
-                            color = FlintTheme.colors.white,
-                            style = FlintTheme.typography.caption1M12,
-                        )
-                    }
-                }
-            }
-
-            Spacer(Modifier.height(24.dp))
-
-            Box(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .height(1.dp)
-                        .background(FlintTheme.colors.gray500),
-            )
-
-            Spacer(Modifier.height(24.dp))
-
-            if (content.isSpoiler) {
-                Spoiler(
-                    onSpoilClick = { onSpoilClick(content.id) },
-                    modifier = Modifier.fillMaxWidth(),
-                ) {
-                    Text(
-                        text = content.reason,
-                        color = FlintTheme.colors.gray100,
-                        style = FlintTheme.typography.body1R16,
-                        modifier =
-                            Modifier
-                                .defaultMinSize(minHeight = 183.dp)
-                                .fillMaxWidth(),
-                    )
-                }
-            } else {
-                Text(
-                    text = content.reason,
-                    color = FlintTheme.colors.gray100,
-                    style = FlintTheme.typography.body1R16,
-                    modifier =
-                        Modifier
-                            .defaultMinSize(minHeight = 183.dp)
-                            .fillMaxWidth(),
-                )
-            }
-
-            Spacer(Modifier.height(64.dp))
-        }
-    }
-}
-
-private data class HeaderPreviewData(
-    val title: String,
-    val isMine: Boolean,
-    val isBookmarked: Boolean,
-)
-
-private class HeaderPreviewProvider : PreviewParameterProvider<HeaderPreviewData> {
-    override val values: Sequence<HeaderPreviewData> =
-        sequenceOf(
-            HeaderPreviewData(
-                title = "한번 보면 못 빠져나오는 여운남는 사랑이야기".repeat(2),
-                isMine = false,
-                isBookmarked = true,
-            ),
-            HeaderPreviewData(
-                title = "한번 보면 못 빠져나오는 여운남는 사랑이야기",
-                isMine = false,
-                isBookmarked = false,
-            ),
-        )
-}
-
-@Preview
-@Composable
-private fun ThumbnailPreview(
-    @PreviewParameter(HeaderPreviewProvider::class) data: HeaderPreviewData,
-) {
-    FlintTheme {
-        Thumbnail(
-            title = data.title,
-            isBookmarked = data.isBookmarked,
-            onSaveDoneButtonClick = {},
-            onSaveNoneButtonClick = {}
-        )
-    }
-}
-
-private data class DescriptionPreviewData(
-    val authorNickname: String,
-    val authorUserRoleType: UserRoleType,
-    val createdAt: String,
-    val collectionContent: String,
-)
-
-private class DescriptionPreviewProvider : PreviewParameterProvider<DescriptionPreviewData> {
-    override val values: Sequence<DescriptionPreviewData> =
-        sequenceOf(
-            DescriptionPreviewData(
-                authorNickname = "키카",
-                authorUserRoleType = UserRoleType.FLINER,
-                createdAt = "2026. 01. 07.",
-                collectionContent = "시간이 흘러도 빛이 바래지 않는,\n사랑의 미묘한 온도를 담은 제 최애 영화 모음집입니다",
-            ),
-            DescriptionPreviewData(
-                authorNickname = "일반유저",
-                authorUserRoleType = UserRoleType.FLING,
-                createdAt = "2026. 01. 15.",
-                collectionContent = "한글자 두글자 세글자 네글자 다섯글자 ".repeat(10),
-            ),
-            DescriptionPreviewData(
-                authorNickname = "관리자",
-                authorUserRoleType = UserRoleType.ADMIN,
-                createdAt = "2026. 01. 01.",
-                collectionContent = "공식 추천 컬렉션입니다",
-            ),
-        )
-}
-
-@Preview
-@Composable
-private fun CollectionDetailDescriptionPreview(
-    @PreviewParameter(DescriptionPreviewProvider::class) data: DescriptionPreviewData,
-) {
-    FlintTheme {
-        CollectionDetailDescription(
-            authorNickname = data.authorNickname,
-            authorUserRoleType = data.authorUserRoleType,
-            createdAt = data.createdAt,
-            collectionContent = data.collectionContent,
-            onAuthorNicknameClick = {}
-        )
-    }
-}
-
-private class ContentPreviewProvider : PreviewParameterProvider<ContentModelNew> {
-    override val values: Sequence<ContentModelNew> =
-        sequenceOf(
-            ContentModelNew(
-                id = "0",
-                title = "드라마 제목",
-                year = 2000,
-                imageUrl = "",
-                director = "가스 제닝스",
-                reason = "달라진 온도\n-\n같은 구도에 채도를 달리해 변해버린 사랑을 시각적으로 담아낸 장면들",
-                isSpoiler = false,
-                isBookmarked = false,
-                bookmarkCount = 42,
-            ),
-            ContentModelNew(
-                id = "0",
-                title = "스포일러 있는 영화",
-                year = 2024,
-                imageUrl = "",
-                director = "감독 이름",
-                reason = "이 내용은 스포일러가 포함되어 있습니다.",
-                isSpoiler = true,
-                isBookmarked = false,
-                bookmarkCount = 42,
-            ),
-            ContentModelNew(
-                id = "0",
-                title = "저장된 영화",
-                year = 2023,
-                imageUrl = "",
-                director = "다른 감독",
-                reason = "내가 저장한 영화입니다.",
-                isSpoiler = false,
-                isBookmarked = true,
-                bookmarkCount = 42,
-            ),
-        )
-}
-
-
-@Preview
-@Composable
-private fun ContentPreview(
-    @PreviewParameter(ContentPreviewProvider::class) content: ContentModelNew,
-) {
-    FlintTheme {
-        Content(
-            content = content,
-            onBookmarkIconClick = {},
-            onSpoilClick = {}
-        )
-    }
-}
-
 private data class ScreenPreviewData(
+    val thumbnailUrl: String,
     val title: String,
     val isBookmarked: Boolean,
     val authorNickname: String,
@@ -964,6 +383,7 @@ private class ScreenPreviewProvider : PreviewParameterProvider<ScreenPreviewData
     override val values: Sequence<ScreenPreviewData> =
         sequenceOf(
             ScreenPreviewData(
+                thumbnailUrl = "",
                 title = "한번 보면 못 빠져나오는 여운남는 사랑이야기",
                 isBookmarked = true,
                 authorNickname = "키카",
@@ -972,6 +392,7 @@ private class ScreenPreviewProvider : PreviewParameterProvider<ScreenPreviewData
                 people = samplePeople,
             ),
             ScreenPreviewData(
+                thumbnailUrl = "https://buly.kr/DEaVFRZ",
                 title = "새로운 컬렉션",
                 isBookmarked = false,
                 authorNickname = "일반유저",
@@ -982,31 +403,33 @@ private class ScreenPreviewProvider : PreviewParameterProvider<ScreenPreviewData
         )
 }
 
-@Preview
-@Composable
-private fun CollectionDetailScreenPreview(
-    @PreviewParameter(ScreenPreviewProvider::class) data: ScreenPreviewData,
-) {
-    FlintTheme {
-        Scaffold { paddingValues: PaddingValues ->
-            CollectionDetailScreen(
-                paddingValues = paddingValues,
-                title = data.title,
-                isBookmarked = data.isBookmarked,
-                authorNickname = data.authorNickname,
-                authorUserRoleType = data.authorUserRoleType,
-                createdAt = "2026. 01. 07.",
-                description = "시간이 흘러도 빛이 바래지 않는,\n사랑의 미묘한 온도를 담은 제 최애 영화 모음집입니다",
-                contents = data.contents,
-                people = data.people,
-                onSaveDoneButtonClick = {},
-                onSaveNoneButtonClick = {},
-                navigateUp = {},
-                onBookmarkIconClick = {},
-                onSpoilClick = {},
-                onAuthorClick = {},
-                onAuthorNicknameClick = {}
-            )
-        }
-    }
-}
+//@Preview
+//@Composable
+//private fun CollectionDetailScreenPreview(
+//    @PreviewParameter(ScreenPreviewProvider::class) data: ScreenPreviewData,
+//) {
+//    FlintTheme {
+//        Scaffold { paddingValues: PaddingValues ->
+//            CollectionDetailScreen(
+//                paddingValues = paddingValues,
+//                title = data.title,
+//                isBookmarked = data.isBookmarked,
+//                authorNickname = data.authorNickname,
+//                authorUserRoleType = data.authorUserRoleType,
+//                createdAt = "2026. 01. 07.",
+//                description = "시간이 흘러도 빛이 바래지 않는,\n사랑의 미묘한 온도를 담은 제 최애 영화 모음집입니다",
+//                contents = data.contents,
+//                people = data.people,
+//                onSaveDoneButtonClick = {},
+//                onSaveNoneButtonClick = {},
+//                navigateUp = {},
+//                onBookmarkIconClick = {},
+//                onSpoilClick = {},
+//                onAuthorClick = {},
+//                onAuthorNicknameClick = {},
+//                onAuthorClick = {},
+//                targetImageUrl = {}
+//            )
+//        }
+//    }
+//}
